# express-infinite-scroll-api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Infinite Scroll Products API",
    "description": "A production-ready RESTful API for managing and retrieving products with infinite scroll capabilities. Features include cursor-based pagination, page-based pagination, comprehensive error handling, request validation, and structured logging with correlation IDs.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api.example.com",
      "description": "Production server"
    }
  ],
  "tags": [
    {
      "name": "Health",
      "description": "Health check and system status endpoints"
    },
    {
      "name": "Products",
      "description": "Product management and retrieval endpoints"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "tags": ["Health"],
        "summary": "Health check endpoint",
        "description": "Returns the health status of the API server and database connectivity. Use this endpoint for monitoring and load balancer health checks.",
        "operationId": "getHealth",
        "responses": {
          "200": {
            "description": "Service is healthy and database is connected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                },
                "example": {
                  "status": "ok",
                  "timestamp": "2025-11-18T10:24:28.583Z",
                  "database": "connected"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable - database connection failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthErrorResponse"
                },
                "example": {
                  "status": "error",
                  "timestamp": "2025-11-18T10:24:28.583Z",
                  "database": "disconnected"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products": {
      "get": {
        "tags": ["Products"],
        "summary": "Get products with cursor-based pagination (NOT IMPLEMENTED YET)",
        "description": "⚠️ **NOT IMPLEMENTED YET** - This endpoint is planned but not yet available. Retrieves a list of products using cursor-based pagination, ideal for infinite scroll implementations. Returns products after the specified cursor ID.",
        "operationId": "getProducts",
        "deprecated": true,
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "description": "Product ID to start fetching after. Omit to start from the beginning.",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            },
            "example": 42
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of products to fetch",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "example": 20
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved products",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CursorPaginatedProducts"
                }
              }
            }
          },
          "400": {
            "description": "Validation error - invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests - rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/paged": {
      "get": {
        "tags": ["Products"],
        "summary": "Get products with page-based pagination and sorting",
        "description": "Retrieves a list of products using traditional page-based pagination with sorting capabilities. Returns products for the specified page number with total count and pagination metadata. Supports sorting by name or price in ascending or descending order.",
        "operationId": "getProductsPaged",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number to retrieve",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "example": 1
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of products per page",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            },
            "example": 50
          },
          {
            "name": "sortBy",
            "in": "query",
            "description": "Field to sort by",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["name", "price"],
              "default": "name"
            },
            "example": "name"
          },
          {
            "name": "sortOrder",
            "in": "query",
            "description": "Sort direction",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["asc", "desc"],
              "default": "asc"
            },
            "example": "asc"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved products",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedProducts"
                }
              }
            }
          },
          "400": {
            "description": "Validation error - invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests - rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Product": {
        "type": "object",
        "required": ["id", "name", "price", "createdAt"],
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique product identifier",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Product name",
            "maxLength": 255,
            "example": "Product 1"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "description": "Product description",
            "example": "Description for product 1"
          },
          "price": {
            "type": "string",
            "description": "Product price in decimal format",
            "pattern": "^\\d+\\.\\d{2}$",
            "example": "366.24"
          },
          "imageUrl": {
            "type": "string",
            "nullable": true,
            "description": "URL to product image",
            "maxLength": 512,
            "example": "https://picsum.photos/seed/1/200"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the product was created",
            "example": "2025-11-18T10:23:55.000Z"
          }
        }
      },
      "CursorPaginatedProducts": {
        "type": "object",
        "required": ["items", "nextCursor"],
        "properties": {
          "items": {
            "type": "array",
            "description": "Array of product objects",
            "items": {
              "$ref": "#/components/schemas/Product"
            }
          },
          "nextCursor": {
            "type": "integer",
            "nullable": true,
            "description": "ID of the next cursor for pagination, null if no more items",
            "example": 21
          }
        }
      },
      "PagedProducts": {
        "type": "object",
        "required": ["items", "hasNextPage", "nextPage", "page", "pageSize", "total"],
        "properties": {
          "items": {
            "type": "array",
            "description": "Array of product objects",
            "items": {
              "$ref": "#/components/schemas/Product"
            }
          },
          "hasNextPage": {
            "type": "boolean",
            "description": "Whether there are more pages available",
            "example": true
          },
          "nextPage": {
            "type": "integer",
            "nullable": true,
            "description": "Next page number, null if no more pages",
            "example": 2
          },
          "page": {
            "type": "integer",
            "description": "Current page number",
            "example": 1
          },
          "pageSize": {
            "type": "integer",
            "description": "Number of items per page",
            "example": 50
          },
          "total": {
            "type": "integer",
            "description": "Total number of products available",
            "example": 1000
          }
        }
      },
      "HealthResponse": {
        "type": "object",
        "required": ["status", "timestamp", "database"],
        "properties": {
          "status": {
            "type": "string",
            "enum": ["ok"],
            "description": "Overall health status",
            "example": "ok"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp of the health check",
            "example": "2025-11-18T10:24:28.583Z"
          },
          "database": {
            "type": "string",
            "enum": ["connected"],
            "description": "Database connection status",
            "example": "connected"
          }
        }
      },
      "HealthErrorResponse": {
        "type": "object",
        "required": ["status", "timestamp", "database"],
        "properties": {
          "status": {
            "type": "string",
            "enum": ["error"],
            "description": "Overall health status",
            "example": "error"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp of the health check",
            "example": "2025-11-18T10:24:28.583Z"
          },
          "database": {
            "type": "string",
            "enum": ["disconnected"],
            "description": "Database connection status",
            "example": "disconnected"
          }
        }
      },
      "ValidationError": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message",
            "example": "Validation error"
          },
          "details": {
            "type": "object",
            "description": "Detailed validation errors from Zod",
            "properties": {
              "formErrors": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "fieldErrors": {
                "type": "object",
                "additionalProperties": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            },
            "example": {
              "formErrors": [],
              "fieldErrors": {
                "limit": ["limit must be an integer between 1 and 100"]
              }
            }
          }
        }
      },
      "RateLimitError": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "string",
            "description": "Rate limit error message",
            "example": "Too many requests from this IP, please try again later"
          }
        }
      },
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message",
            "example": "Internal server error"
          },
          "stack": {
            "type": "string",
            "description": "Stack trace (development only)",
            "example": "Error: Something went wrong\\n    at ..."
          }
        }
      }
    },
    "securitySchemes": {},
    "headers": {
      "X-RateLimit-Limit": {
        "description": "The maximum number of requests allowed in the current time window",
        "schema": {
          "type": "integer",
          "example": 100
        }
      },
      "X-RateLimit-Remaining": {
        "description": "The number of requests remaining in the current time window",
        "schema": {
          "type": "integer",
          "example": 99
        }
      },
      "X-RateLimit-Reset": {
        "description": "The time at which the current rate limit window resets (Unix timestamp)",
        "schema": {
          "type": "integer",
          "example": 1700000000
        }
      }
    }
  },
  "security": [],
  "x-rate-limit": {
    "production": "100 requests per 15 minutes",
    "development": "1000 requests per 15 minutes"
  },
  "x-features": {
    "logging": "Pino structured logging with correlation IDs",
    "security": "Helmet.js, CORS whitelist, rate limiting, request size limits",
    "validation": "Zod schema validation",
    "compression": "Gzip compression enabled"
  }
}